Crate rusoto_amplify

Amplify is a fully managed continuous deployment and hosting service for modern web apps.

If you're using the service, you're probably looking for AmplifyClient and Amplify.

Structs

AmplifyClient
A client for the Amplify API.

App
Amplify App represents different branches of a repository for building, deploying, and hosting.

Branch
Branch for an Amplify App, which maps to a 3rd party repository branch.

CreateAppRequest
Request structure used to create Apps in Amplify.

CreateAppResult

CreateBranchRequest
Request structure for a branch create request.

CreateBranchResult
Result structure for create branch request.

CreateDomainAssociationRequest
Request structure for create Domain Association request.

CreateDomainAssociationResult
Result structure for the create Domain Association request.

CustomRule
Custom rewrite / redirect rule.

DeleteAppRequest
Request structure for an Amplify App delete request.

DeleteAppResult
Result structure for an Amplify App delete request.

DeleteBranchRequest
Request structure for delete branch request.

DeleteBranchResult
Result structure for delete branch request.

DeleteDomainAssociationRequest
Request structure for the delete Domain Association request.

DeleteDomainAssociationResult

DeleteJobRequest
Request structure for delete job request.

DeleteJobResult
Result structure for the delete job request.

DomainAssociation
Structure for Domain Association, which associates a custom domain with an Amplify App.

GetAppRequest
Request structure for get App request.

GetAppResult

GetBranchRequest
Result structure for get branch request.

GetBranchResult

GetDomainAssociationRequest
Request structure for the get Domain Association request.

GetDomainAssociationResult
Result structure for the get Domain Association request.

GetJobRequest
Request structure for get job request.

GetJobResult

Job
Structure for an execution job for an Amplify App.

JobSummary
Structure for the summary of a Job.

ListAppsRequest
Request structure for an Amplify App list request.

ListAppsResult
Result structure for an Amplify App list request.

ListBranchesRequest
Request structure for list branches request.

ListBranchesResult
Result structure for list branches request.

ListDomainAssociationsRequest
Request structure for the list Domain Associations request.

ListDomainAssociationsResult
Result structure for the list Domain Association request.

ListJobsRequest
Request structure for list job request.

ListJobsResult
Maximum number of records to list in a single response.

ProductionBranch
Structure with Production Branch information.

StartJobRequest
Request structure for Start job request.

StartJobResult
Result structure for run job request.

Step
Structure for an execution step for an execution job, for an Amplify App.

StopJobRequest
Request structure for stop job request.

StopJobResult
Result structure for the stop job request.

SubDomain
Subdomain for the Domain Association.

SubDomainSetting
Setting for the Subdomain.

UpdateAppRequest
Request structure for update App request.

UpdateAppResult
Result structure for an Amplify App update request.

UpdateBranchRequest
Request structure for update branch request.

UpdateBranchResult
Result structure for update branch request.

UpdateDomainAssociationRequest
Request structure for update Domain Association request.

UpdateDomainAssociationResult
Result structure for the update Domain Association request.

Enums

CreateAppError
Errors returned by CreateApp

CreateBranchError
Errors returned by CreateBranch

CreateDomainAssociationError
Errors returned by CreateDomainAssociation

DeleteAppError
Errors returned by DeleteApp

DeleteBranchError
Errors returned by DeleteBranch

DeleteDomainAssociationError
Errors returned by DeleteDomainAssociation

DeleteJobError
Errors returned by DeleteJob

GetAppError
Errors returned by GetApp

GetBranchError
Errors returned by GetBranch

GetDomainAssociationError
Errors returned by GetDomainAssociation

GetJobError
Errors returned by GetJob

ListAppsError
Errors returned by ListApps

ListBranchesError
Errors returned by ListBranches

ListDomainAssociationsError
Errors returned by ListDomainAssociations

ListJobsError
Errors returned by ListJobs

StartJobError
Errors returned by StartJob

StopJobError
Errors returned by StopJob

UpdateAppError
Errors returned by UpdateApp

UpdateBranchError
Errors returned by UpdateBranch

UpdateDomainAssociationError
Errors returned by UpdateDomainAssociation

Traits

Amplify
Trait representing the capabilities of the Amplify API. Amplify clients implement this trait.