Struct rusoto_amplify::AmplifyClient

pub struct AmplifyClient { /* fields omitted */ }

A client for the Amplify API.

Methods

impl AmplifyClient

pub fn new(region: Region) -> AmplifyClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AmplifyClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,