Trait rusoto_amplify::Amplify

pub trait Amplify {
    fn create_app(
        &self, 
        input: CreateAppRequest
    ) -> RusotoFuture<CreateAppResult, CreateAppError>;

    fn create_branch(
        &self, 
        input: CreateBranchRequest
    ) -> RusotoFuture<CreateBranchResult, CreateBranchError>;

    fn create_domain_association(
        &self, 
        input: CreateDomainAssociationRequest
    ) -> RusotoFuture<CreateDomainAssociationResult, CreateDomainAssociationError>;

    fn delete_app(
        &self, 
        input: DeleteAppRequest
    ) -> RusotoFuture<DeleteAppResult, DeleteAppError>;

    fn delete_branch(
        &self, 
        input: DeleteBranchRequest
    ) -> RusotoFuture<DeleteBranchResult, DeleteBranchError>;

    fn delete_domain_association(
        &self, 
        input: DeleteDomainAssociationRequest
    ) -> RusotoFuture<DeleteDomainAssociationResult, DeleteDomainAssociationError>;

    fn delete_job(
        &self, 
        input: DeleteJobRequest
    ) -> RusotoFuture<DeleteJobResult, DeleteJobError>;

    fn get_app(
        &self, 
        input: GetAppRequest
    ) -> RusotoFuture<GetAppResult, GetAppError>;

    fn get_branch(
        &self, 
        input: GetBranchRequest
    ) -> RusotoFuture<GetBranchResult, GetBranchError>;

    fn get_domain_association(
        &self, 
        input: GetDomainAssociationRequest
    ) -> RusotoFuture<GetDomainAssociationResult, GetDomainAssociationError>;

    fn get_job(
        &self, 
        input: GetJobRequest
    ) -> RusotoFuture<GetJobResult, GetJobError>;

    fn list_apps(
        &self, 
        input: ListAppsRequest
    ) -> RusotoFuture<ListAppsResult, ListAppsError>;

    fn list_branches(
        &self, 
        input: ListBranchesRequest
    ) -> RusotoFuture<ListBranchesResult, ListBranchesError>;

    fn list_domain_associations(
        &self, 
        input: ListDomainAssociationsRequest
    ) -> RusotoFuture<ListDomainAssociationsResult, ListDomainAssociationsError>;

    fn list_jobs(
        &self, 
        input: ListJobsRequest
    ) -> RusotoFuture<ListJobsResult, ListJobsError>;

    fn start_job(
        &self, 
        input: StartJobRequest
    ) -> RusotoFuture<StartJobResult, StartJobError>;

    fn stop_job(
        &self, 
        input: StopJobRequest
    ) -> RusotoFuture<StopJobResult, StopJobError>;

    fn update_app(
        &self, 
        input: UpdateAppRequest
    ) -> RusotoFuture<UpdateAppResult, UpdateAppError>;

    fn update_branch(
        &self, 
        input: UpdateBranchRequest
    ) -> RusotoFuture<UpdateBranchResult, UpdateBranchError>;

    fn update_domain_association(
        &self, 
        input: UpdateDomainAssociationRequest
    ) -> RusotoFuture<UpdateDomainAssociationResult, UpdateDomainAssociationError>;
}

Trait representing the capabilities of the Amplify API. Amplify clients implement this trait.

Required methods

fn create_app(
&self,
input: CreateAppRequest
) -> RusotoFuture<CreateAppResult, CreateAppError>

Creates a new Amplify App.

fn create_branch(
&self,
input: CreateBranchRequest
) -> RusotoFuture<CreateBranchResult, CreateBranchError>

Creates a new Branch for an Amplify App.

fn create_domain_association(
&self,
input: CreateDomainAssociationRequest
) -> RusotoFuture<CreateDomainAssociationResult, CreateDomainAssociationError>

Create a new DomainAssociation on an App

fn delete_app(
&self,
input: DeleteAppRequest
) -> RusotoFuture<DeleteAppResult, DeleteAppError>

Delete an existing Amplify App by appId.

fn delete_branch(
&self,
input: DeleteBranchRequest
) -> RusotoFuture<DeleteBranchResult, DeleteBranchError>

Deletes a branch for an Amplify App.

fn delete_domain_association(
&self,
input: DeleteDomainAssociationRequest
) -> RusotoFuture<DeleteDomainAssociationResult, DeleteDomainAssociationError>

Deletes a DomainAssociation.

fn delete_job(
&self,
input: DeleteJobRequest
) -> RusotoFuture<DeleteJobResult, DeleteJobError>

Delete a job, for an Amplify branch, part of Amplify App.

fn get_app(
&self,
input: GetAppRequest
) -> RusotoFuture<GetAppResult, GetAppError>

Retrieves an existing Amplify App by appId.

fn get_branch(
&self,
input: GetBranchRequest
) -> RusotoFuture<GetBranchResult, GetBranchError>

Retrieves a branch for an Amplify App.

fn get_domain_association(
&self,
input: GetDomainAssociationRequest
) -> RusotoFuture<GetDomainAssociationResult, GetDomainAssociationError>

Retrieves domain info that corresponds to an appId and domainName.

fn get_job(
&self,
input: GetJobRequest
) -> RusotoFuture<GetJobResult, GetJobError>

Get a job for a branch, part of an Amplify App.

fn list_apps(
&self,
input: ListAppsRequest
) -> RusotoFuture<ListAppsResult, ListAppsError>

Lists existing Amplify Apps.

fn list_branches(
&self,
input: ListBranchesRequest
) -> RusotoFuture<ListBranchesResult, ListBranchesError>

Lists branches for an Amplify App.

fn list_domain_associations(
&self,
input: ListDomainAssociationsRequest
) -> RusotoFuture<ListDomainAssociationsResult, ListDomainAssociationsError>

List domains with an app

fn list_jobs(
&self,
input: ListJobsRequest
) -> RusotoFuture<ListJobsResult, ListJobsError>

List Jobs for a branch, part of an Amplify App.

fn start_job(
&self,
input: StartJobRequest
) -> RusotoFuture<StartJobResult, StartJobError>

Starts a new job for a branch, part of an Amplify App.

fn stop_job(
&self,
input: StopJobRequest
) -> RusotoFuture<StopJobResult, StopJobError>

Stop a job that is in progress, for an Amplify branch, part of Amplify App.

fn update_app(
&self,
input: UpdateAppRequest
) -> RusotoFuture<UpdateAppResult, UpdateAppError>

Updates an existing Amplify App.

fn update_branch(
&self,
input: UpdateBranchRequest
) -> RusotoFuture<UpdateBranchResult, UpdateBranchError>

Updates a branch for an Amplify App.

fn update_domain_association(
&self,
input: UpdateDomainAssociationRequest
) -> RusotoFuture<UpdateDomainAssociationResult, UpdateDomainAssociationError>

Create a new DomainAssociation on an App

Loading content...